test: force tox to run pip with --upgrade
This forces it to properly install the local module since it thinks it
is "installed" locally already in the repo probably due to an egg dist
dir created during the sdist phase.

Also, fix paths for coverage so they are local to the module install
path instead of the repo dir. This should fix issues coveralls has with
finding the related files from github to show line/branch coverage.