bind: handle mounting destinations that are symlinks
Previously they would resolve to locations outside the chroot and
therefore wouldn't work properly within the chroot context.

This method also assumes that if the destination pointed to be the
symlink doesn't exist it'll just create it. We can consider making this
an error or warning condition later if we want to be more strict.