add option to skip changing to the newroot after chrooting
This adds the option to the context manager and provides the option flag
'--skip-chdir' for the cli tool.

This is similar to the option for chroot(1) but this version also allows
skipping the directory change when the new root isn't '/'. In other
words, you can use a new chroot environment on the current host system
rootfs.

This can be handy in certain cases, e.g. if the chroot env contains
tools that you want to use on the main rootfs and you don't want to
manipulation PATH or other things. Extending this idea, you could have a
number of chroots with a bunch of different tool configurations that you
want to use against the data in the main rootfs.

Another situation could be that the chroots are extremely lightweight
only providing files required to bring a service or two they bundle up
while the main rootfs contains the data they should operate on.