add initial user namespace support
Note that this requires a network namespace to be used as well which
currently means network routing out to the regular network won't work
under the default setup that is brought up since we only bring up a
loopback device if iproute2 is installed in the chroot.

In other words, use root privs if you need external network access for
now.

Fixes #13.